Remove utils.streamlines. Lots of moving
